fix unsignedness in dbus.c

+++ sflphone-gtk/src/dbus.c
@@ -1266,7 +1266,7 @@ guint
 dbus_get_max_calls( void )
 {
 	GError* error = NULL;
-	guint calls;
+	gint calls;
 	org_sflphone_SFLphone_ConfigurationManager_get_max_calls(
 			configurationManagerProxy,
 			&calls,
@@ -1276,7 +1276,7 @@ dbus_get_max_calls( void )
 		g_error_free(error);
 	}
 	g_print("GET MAX CALLS = %i\n" , calls);
-	return calls;
+	return (guint)calls;
 }
 
 void
@@ -1365,7 +1365,7 @@ guint
 dbus_get_notify( void )
 {
   g_print("Before dbus_get_notif_level()\n");
-	guint level;
+	gint level;
 	GError* error = NULL;
 	org_sflphone_SFLphone_ConfigurationManager_get_notify(
 			configurationManagerProxy,
@@ -1379,7 +1379,7 @@ dbus_get_notify( void )
 	else
 	  g_print("Called dbus_get_notif_level\n");
 	
-	return level;
+	return (guint)level;
 }
 
 void
@@ -1401,7 +1401,7 @@ guint
 dbus_get_mail_notify( void )
 {
   g_print("Before dbus_get_mail_notif_level()\n");
-	guint level;
+	gint level;
 	GError* error = NULL;
 	org_sflphone_SFLphone_ConfigurationManager_get_mail_notify(
 			configurationManagerProxy,
@@ -1415,7 +1415,7 @@ dbus_get_mail_notify( void )
 	else
 	  g_print("Called dbus_get_mail_notif_level\n");
 	
-	return level;
+	return (guint)level;
 }
 
 void
